Parser Token To Node
